Make sure all tests are run with TomEE 7
Rename tests and update dependencies to make all tests execute.

Use a recent Arquillian version and TomEE 7.
