+++ vaadin-cdi/pom.xml
@@ -13,12 +13,13 @@
         
         <!-- In addition to POMs, these version numbers are in ArchiveProvider for tests -->
         <vaadin.version>7.5.1</vaadin.version>
-        <deltaspike.version>1.4.1</deltaspike.version>
+        <deltaspike.version>1.7.2</deltaspike.version>
 
         <!-- Testing dependencies -->
-        <selenium.version>2.43.1</selenium.version>
-        <shiro.version>1.2.3</shiro.version>
-        <arquillian.version>1.1.7.Final</arquillian.version>
+        <selenium.version>2.47.1</selenium.version>
+        <!-- This needs to be in sync with the class ShiroTest -->
+        <shiro.version>1.3.2</shiro.version>
+        <arquillian.version>1.1.10.Final</arquillian.version>
 
         <!-- ZIP Manifest fields -->
         <Implementation-Version>${project.version}</Implementation-Version>
@@ -178,24 +179,10 @@
             <dependencies>
                 <!-- Embedded TomEE -->
                 <dependency>
-                    <groupId>org.apache.openejb</groupId>
-                    <artifactId>apache-tomee</artifactId>
-                    <version>1.6.0.2</version>
-                    <classifier>plus</classifier>
-                    <type>zip</type>
-                    <scope>test</scope>
-                </dependency>
-                <dependency>
-                    <groupId>org.apache.openejb</groupId>
+                    <groupId>org.apache.tomee</groupId>
                     <artifactId>arquillian-tomee-embedded</artifactId>
-                    <version>1.6.0.2</version>
+                    <version>7.0.2</version>
                     <scope>test</scope>
-                </dependency>
-                <dependency>
-                  <groupId>javax.servlet</groupId>
-                  <artifactId>javax.servlet-api</artifactId>
-                  <version>3.1.0</version>
-                  <scope>provided</scope>
                 </dependency>
                 <!-- Java EE 6 API. Must be last to be overridden by tomee for
                      the test phase -->
@@ -254,6 +241,13 @@
             <version>${deltaspike.version}</version>
             <scope>runtime</scope>
         </dependency>
+        
+        <dependency>
+           <groupId>javax.enterprise</groupId>
+           <artifactId>cdi-api</artifactId>
+           <version>1.2</version>
+           <scope>provided</scope>
+        </dependency>
 
         <!-- Test dependencies -->
         <!-- Arquillian -->
@@ -272,12 +266,12 @@
             <groupId>org.jboss.arquillian.graphene</groupId>
             <artifactId>arquillian-graphene</artifactId>
             <type>pom</type>
-            <version>2.0.3.Final</version>
+            <version>2.1.0.Alpha3</version>
             <scope>test</scope>
             <exclusions>
                 <exclusion>
-                    <groupId>org.mortbay.jetty</groupId>
-                    <artifactId>servlet-api-2.5</artifactId>
+                     <groupId>javax.enterprise</groupId>
+                     <artifactId>cdi-api</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
@@ -285,7 +279,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.11</version>
+            <version>4.12</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -365,23 +359,21 @@
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <version>2.13</version>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.19.1</version>
                 <configuration>
-                   <!-- Assertions must be disabled when @UIScoped is a @NormalScope
+                    <!-- Assertions must be disabled when @UIScoped is a @NormalScope
                         otherwise the core framework will fail assertions that compare
                         a direct reference to a proxy, such as in VaadinSession#removeUI
                         and ConnectorTracker#cleanConnectorMap (#14508) -->
-                   <enableAssertions>false</enableAssertions>
+                    <enableAssertions>false</enableAssertions>
+                    <!-- Fork every test because it will launch a separate AS instance -->
+                    <forkMode>always</forkMode>
+                    <systemPropertyVariables>
+                       <javax.el.ExpressionFactory>org.apache.el.ExpressionFactoryImpl</javax.el.ExpressionFactory>
+                    </systemPropertyVariables>
+                    <redirectTestOutputToFile>false</redirectTestOutputToFile>
                 </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>integration-test</goal>
-                            <goal>verify</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
             <plugin>
                 <artifactId>maven-deploy-plugin</artifactId>
