+++ vaadin-cdi/src/test/java/com/vaadin/cdi/CDIIntegrationWithCustomDeploymentTest.java
@@ -16,7 +16,6 @@
 
 package com.vaadin.cdi;
 
-import static com.vaadin.cdi.internal.Conventions.deriveMappingForUI;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -28,32 +27,27 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Before;
 import org.junit.Test;
 
-import com.vaadin.cdi.uis.PlainColidingAlternativeUI;
-import com.vaadin.cdi.uis.PlainUI;
+import com.vaadin.cdi.uis.CustomMappingUI;
 
-public class CDIIntegrationWithConflictingDeployment extends
+public class CDIIntegrationWithCustomDeploymentTest extends
         AbstractManagedCDIIntegrationTest {
 
     @Before
     public void resetCounter() {
-        PlainUI.resetCounter();
-        PlainColidingAlternativeUI.resetCounter();
+        CustomMappingUI.resetCounter();
     }
 
-    @Deployment(name = "alternativeUiPathCollision")
-    public static WebArchive alternativeAndActiveWithSamePath() {
-        return ArchiveProvider.createWebArchive("alternativeUiPathCollision",
-                PlainUI.class, PlainColidingAlternativeUI.class);
+    @Deployment(name = "customURIMapping")
+    public static WebArchive archiveWithCustomURIMapping() {
+        return ArchiveProvider
+                .createWebArchive("custom", CustomMappingUI.class);
     }
 
     @Test
-    @OperateOnDeployment("alternativeUiPathCollision")
-    public void alternativeDoesNotColideWithPath() throws MalformedURLException {
-        final String plainUIPath = deriveMappingForUI(PlainUI.class);
-        final String plainAlternativeUI = deriveMappingForUI(PlainColidingAlternativeUI.class);
-        assertThat(plainUIPath, is(plainAlternativeUI));
-        openWindow(plainUIPath);
-        assertThat(PlainUI.getNumberOfInstances(), is(1));
-        assertThat(PlainColidingAlternativeUI.getNumberOfInstances(), is(0));
+    @OperateOnDeployment("customURIMapping")
+    public void customServletMapping() throws MalformedURLException {
+        assertThat(CustomMappingUI.getNumberOfInstances(), is(0));
+        openWindow("customURI/");
+        assertThat(CustomMappingUI.getNumberOfInstances(), is(1));
     }
 }
