+++ vaadin-cdi/src/test/java/com/vaadin/cdi/DeploymentTestSuiteIT.java
@@ -7,17 +7,17 @@ import com.vaadin.cdi.internal.ConventionsTest;
 import com.vaadin.cdi.shiro.ShiroTest;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({ CDIIntegrationWithCustomDeployment.class,
-        CDIUIProviderTest.class, ConventionsTest.class, MultipleRootUIs.class,
-        CDIIntegrationWithConflictingDeployment.class,
-        CDIIntegrationWithConflictingUIPath.class,
-        CDIIntegrationWithDefaultDeployment.class, RootViewAtContextRoot.class,
-        MultipleAccessIsolation.class, ScopedInstances.class,
+@Suite.SuiteClasses({ CDIIntegrationWithCustomDeploymentTest.class,
+        CDIUIProviderTest.class, ConventionsTest.class, MultipleRootUIsTest.class,
+        CDIIntegrationWithConflictingDeploymentTest.class,
+        CDIIntegrationWithConflictingUIPathTest.class,
+        CDIIntegrationWithDefaultDeploymentTest.class, RootViewAtContextRootTest.class,
+        MultipleAccessIsolationTest.class, ScopedInstancesTest.class,
         InjectionTest.class, ConsistentInjectionTest.class,
-        QualifiedInjection.class, MultipleSessionTest.class,
-        ScopedProducer.class, CrossInjection.class, ShiroTest.class,
-        InappropriateNestedServletInDeployment.class,
-        InappropriateCDIViewInDeployment.class, NonPassivatingBeanTest.class })
+        QualifiedInjectionTest.class, MultipleSessionTest.class,
+        ScopedProducerTest.class, CrossInjectionTest.class, ShiroTest.class,
+        InappropriateNestedServletInDeploymentTest.class,
+        InappropriateCDIViewInDeploymentTest.class, NonPassivatingBeanTest.class })
 public class DeploymentTestSuiteIT {
 
 }
