+++ vaadin-cdi/src/test/java/com/vaadin/cdi/InappropriateCDIViewInDeploymentTest.java
@@ -1,7 +1,5 @@
 package com.vaadin.cdi;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
 import java.net.MalformedURLException;
@@ -10,26 +8,17 @@ import org.jboss.arquillian.container.spi.client.container.DeploymentException;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.InSequence;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.Before;
 import org.junit.Test;
 
-import com.vaadin.cdi.uis.AnotherPathCollisionUI;
-import com.vaadin.cdi.uis.PathCollisionUI;
-import com.vaadin.cdi.uis.RootUI;
+import com.vaadin.cdi.views.CDIViewNotImplementingView;
 
-public class CDIIntegrationWithConflictingUIPath extends
+public class InappropriateCDIViewInDeploymentTest extends
         AbstractCDIIntegrationTest {
 
-    @Before
-    public void resetCounter() {
-        PathCollisionUI.resetCounter();
-        AnotherPathCollisionUI.resetCounter();
-    }
-
-    @Deployment(name = "uiPathCollision", managed = false)
+    @Deployment(name = "cdiViewWithoutView", managed = false)
     public static WebArchive multipleUIsWithSamePath() {
-        return ArchiveProvider.createWebArchive("uiPathCollision",
-                PathCollisionUI.class, AnotherPathCollisionUI.class);
+        return ArchiveProvider.createWebArchive("cdiViewWithoutView",
+                CDIViewNotImplementingView.class);
     }
 
     /**
@@ -38,19 +27,20 @@ public class CDIIntegrationWithConflictingUIPath extends
      */
     @Test
     @InSequence(-2)
-    public void uiPathCollisionBreaksDeployment() throws MalformedURLException {
-        assertThat(RootUI.getNumberOfInstances(), is(0));
+    public void cdiViewWithoutViewBreaksDeployment()
+            throws MalformedURLException {
         try {
             System.out.println("DEPLOYING");
             // Deployment doesn't declare that it can throw DeploymentException
             // De-facto it can
-            deployer.deploy("uiPathCollision");
+            deployer.deploy("cdiViewWithoutView");
             System.out.println("Deployed");
-            fail("Duplicate deployment paths should not be deployable");
+            fail("CDIView that does not implement View should not be deployable");
             throw new DeploymentException(null);
         } catch (DeploymentException e) {
             // Correct response
             System.out.println("Exiting try block");
         }
     }
+
 }
