+++ vaadin-cdi/src/test/java/com/vaadin/cdi/MultipleRootUIsTest.java
@@ -22,40 +22,44 @@ import java.net.MalformedURLException;
 
 import org.jboss.arquillian.container.spi.client.container.DeploymentException;
 import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit.InSequence;
+import org.jboss.arquillian.container.test.api.RunAsClient;
+import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
-import com.vaadin.cdi.uis.UIWithNestedServlet;
+import com.vaadin.cdi.uis.CustomMappingUI;
+import com.vaadin.cdi.uis.RootUI;
 
-public class InappropriateNestedServletInDeployment extends
-        AbstractCDIIntegrationTest {
+@RunAsClient
+@RunWith(Arquillian.class)
+public class MultipleRootUIsTest extends AbstractCDIIntegrationTest {
 
-    @Deployment(name = "nestedServlet", managed = false)
-    public static WebArchive alternativeAndActiveWithSamePath() {
-        return ArchiveProvider.createWebArchive("nestedServlet",
-                UIWithNestedServlet.class);
+    @Before
+    public void resetCounter() {
+        RootUI.resetCounter();
+        CustomMappingUI.resetCounter();
+    }
+
+    @Deployment(name = "multipleRoots", managed = false)
+    public static WebArchive archiveWithMultipleRoots() {
+        return ArchiveProvider.createWebArchive("multipleRoots", RootUI.class,
+                CustomMappingUI.class);
     }
 
     /**
-     * Tests invalid deployment nested servlet within a UI class. Should be
-     * started first -- Arquillian deployments are not perfectly isolated.
+     * Tests invalid deployment of multiple roots within a WAR Should be before
+     * the regular tests -- Arquillian deployments are not perfectly isolated.
      */
     @Test
-    @InSequence(-2)
-    public void nestedServletBreaksDeployment() throws MalformedURLException {
+    public void multipleRootsBreakDeployment() throws MalformedURLException {
         try {
-            System.out.println("DEPLOYING");
-            // Deployment doesn't declare that it can throw DeploymentException
-            // De-facto it can
-            deployer.deploy("nestedServlet");
-            System.out.println("Deployed");
-            fail("Servlet class nested in the UI should not be deployable");
+            deployer.deploy("multipleRoots");
+            fail("Multiple roots should not be deployable");
             throw new DeploymentException(null);
         } catch (DeploymentException e) {
             // Correct response
-            System.out.println("Exiting try block");
         }
     }
-
 }
